Expanding Banner recipe and color tokens, refactor theme location
